src/project/requirements/development.txt
-r _base.txt
python-dotenv==0.21.0
pylint==2.15.10
black==22.12.0
pylint-django==2.5.3